# coco3: video optimisations